Refactor theme management: consolidate theme creation in App component, add global styles for focus rings, and remove obsolete theme index file.
